+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased](https://github.com/python-social-auth/social-app-pyramid/commits/master)
 
+### Added
+- Added partial pipeline db storage solution
+
 ### Changed
 - Allow apps to have any type of user_id field (port of [#1040](https://github.com/omab/python-social-auth/pull/1040)
   by prmtl)